+++ src/services/productService.js
@@ -1,16 +1,26 @@
-const { productDao } = require("../models/productDao");
-//const { throwError } = require("../utils/throwError");
-const showMain = async (req, res) => {
-  // query thread from db
-  const products = await productDao.showMain(req, res);
+const productDao = require("../models/productDao");
 
-  return products;
-};
+const showMain = (req, res) => {
+  /*
+    // 비밀번호 number, 소문자, 대문자, 특수문자, 5자 이상
+    const passwordRegex = /^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*\W)(?!.*).$/
 
-const showSpecificProduct = async (req, res) => {
-  let productId = req.productId;
+    if (!passwordRegex.test(password)) {
+        const error = new Error ("NO_CHARACTERS")
+        error.status = 400
+        throw error
+    }
+*/
 };
-module.exports = {
-  showMain,
-  showSpecificProduct,
+const showSpecificProduct = (req, res) => {
+  /*
+    // 비밀번호 number, 소문자, 대문자, 특수문자, 5자 이상
+    const passwordRegex = /^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*\W)(?!.*).$/
+
+    if (!passwordRegex.test(password)) {
+        const error = new Error ("NO_CHARACTERS")
+        error.status = 400
+        throw error
+    }
+*/
 };
